requirements.txt
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.7.0
filelock==3.9.0
fonttools==4.50.0
fsspec==2023.4.0
gitdb==4.0.11
GitPython==3.1.42
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
liac-arff==2.5.0
lightning-utilities==0.11.0
MarkupSafe==2.1.3
matplotlib==3.8.3
minio==7.2.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
openml==0.14.2
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.2
pycparser==2.21
pycryptodome==3.20.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
sentry-sdk==1.43.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
threadpoolctl==3.4.0
torch==2.2.1+cu121
torchaudio==2.2.1+cu121
torcheval==0.0.7
torchmetrics==1.3.2
torchvision==0.17.1+cu121
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.4
xmltodict==0.13.0
